+++ src/main/java/com/maxmind/geoip2/model/AbstractResponse.java
@@ -3,6 +3,7 @@ package com.maxmind.geoip2.model;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.databind.MapperFeature;
 import com.fasterxml.jackson.databind.json.JsonMapper;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import java.io.IOException;
 
 /**
@@ -18,6 +19,7 @@ public abstract class AbstractResponse {
     public String toJson() throws IOException {
         JsonMapper mapper = JsonMapper.builder()
             .disable(MapperFeature.CAN_OVERRIDE_ACCESS_MODIFIERS)
+            .addModule(new JavaTimeModule())   
             .serializationInclusion(Include.NON_NULL)
             .serializationInclusion(Include.NON_EMPTY)
             .build();